start for the Statsview redesign
