+++ emotion_tracker/src/components/TabletEmotionButtons.js
@@ -1,32 +1,77 @@
 import EmotionStats from "./EmotionStats";
 import EmotionStatsDay from "./EmotionStats";
 import React, { useState, useEffect } from "react";
-import './EmotionButtons.css';
-import emotionData from "./data/emotionData";
+import '../EmotionButtons.css';
+
+
+const buttonData = [
+  {
+    label: "Happy",
+    id: "1",
+    icon: "sentiment_satisfied",
+    disabled: false
+  },
+  {
+    label: "Angry",
+    id: "2",
+    icon: "sentiment_extremely_dissatisfied",
+    disabled: false
+  },
+  {
+    label: "Scared",
+    id: "3",
+    icon: "mood_bad",
+    disabled: false
+  },
+
+  {
+    label: "Excited",
+    id: "5",
+    icon: "sentiment_very_satisfied",
+    disabled: false
+  },
+  {
+    label: "Sad",
+    id: "4",
+    icon: "sentiment_dissatisfied",
+    disabled: false
+  },
+  {
+    label: "Neutral",
+    id: "6",
+    icon: "sentiment_neutral",
+    disabled: false
+  },
+];
+
+
+
 
 const getButtonClassName = (label) => {
   // Generate a unique class name based on the button's label
   return `${label}`;
 };
 
-function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
+
+
+function TabletEmotionButton({ updateStats, setViewCondition, viewCondition }) {
   const [statsData, setStatsData] = useState();
   const [statsTodayData, setStatsTodayData] = useState();
   const [update, setUpdate] = useState(false);
-  const [timerText, setTimerText] = useState("");
+  const [timerText, setTimerText] = useState("")
   const [buttonActive, setButtonActive] = useState(true);
   const [time, setTime] = useState(0);
-  const timerTimeMs = 15000;
+  const timerTimeMs = 1000;
   const [startAnimation, setStartAnimation] = useState(false);
-  const [clicked, setClicked] = useState(0);
+  const [clicked, setClicked] = useState(0)
 
   // TIMER
   const timerStart = (e) => {
     console.log("timerStart ");
     e.preventDefault();
-    setButtonActive(false);
-    const now = Date.now();
-    localStorage.setItem("timer", now);
+      setButtonActive(false);
+      const nyt = Date.now()
+      localStorage.setItem('timer', nyt);
   };
 
   const timerTick = () => {
@@ -36,9 +81,9 @@ function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
       setTime((timerTimeMs-res) > 0 ? timerTimeMs-res : 0);
       // console.log(res);
 
-      if (res > timerTimeMs) {
+      if (res > timerTimeMs){
         setButtonActive(true);
-        setClicked(0);
+        setClicked(0)
       } else {
         setButtonActive(false);
       }
@@ -52,10 +97,11 @@ function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
     return () => clearInterval(timer);
   }, []);
 
-  // END OF TIMER
+// END OF TIMER
 
-  // post emotion to database
+
   const addEmotion = async (id) => {
+
     try {
       const response = await fetch("http://localhost:3001/add/addemotion", {
         method: "POST",
@@ -67,40 +113,37 @@ function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
       if (!response.ok) {
         throw new Error("Error adding emotion");
       }
-      setUpdate(!update);
+      // handle success response here
+      setUpdate(!update)
+      // setStatsData()
+      // setStatsTodayData("kku")
     } catch (error) {
       // handle error here
-      console.error(error);
     }
   };
 
+
   const buttonClicked = async (id, e) => {
+    // button clicked
+
+    console.log("button clicked " + id);
     addEmotion(id);
     setClicked(id);
     timerStart(e);
-  };
+
+      }
 
   // const [disable, setDisable] = useState(false)
 
   return (
     <div className="content">
       <div className="emotion-buttons">
-        {emotionData.map((button) => (
+        {buttonData.map((button) => (
           <button
-            style={{
-              animation: startAnimation ? "fadeIn 3s, forwards" : "none",
-            }}
+            style={{animation: startAnimation ? "fadeIn 3s, forwards" : "none"}}
             key={button.label}
-            className={
-              clicked !== button.id && !buttonActive
-                ? getButtonClassName(button.label + "-disabled")
-                : getButtonClassName(button.label)
-            }
-            id={
-              clicked === button.id
-                ? getButtonClassName(button.label + "-clicked")
-                : getButtonClassName(button.label)
-            }
+            className={clicked !== button.id && !buttonActive ? getButtonClassName(button.label + "-disabled") : getButtonClassName(button.label)}
+            id = {clicked === button.id ? getButtonClassName(button.label + "-clicked") : getButtonClassName(button.label)}
             disabled={!buttonActive}
             onClick={(e) => buttonClicked(button.id, e)}
           >
@@ -110,13 +153,11 @@ function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
             </div>
           </button>
         ))}
-      </div>
-      <div style={{ visibility: buttonActive ? "hidden" : "visible" }}>
-        <p className="infoText">
-          Share your feelings again in {Math.floor(time / 1000 / 60)} mins,{" "}
-          {Math.floor((time / 1000) % 60)} secs
-        </p>
-      </div>
+            </div>
+            <div style={{visibility: buttonActive ? "hidden" : "visible"}}>
+          <p class="infoText">
+          </p>
+          </div>
       <EmotionStats
         // statsData={statsData}
         // statsTodayData={statsTodayData}
@@ -125,6 +166,6 @@ function EmotionButton({ updateStats, setViewCondition, viewCondition }) {
     </div>
   );
 }
-export default EmotionButton;
+export default TabletEmotionButton;
 // build test
