+++ emotion_tracker/src/components/Timer.js
