+++ emotion_tracker/src/views/TooFarAwayView.js
